Blueprint/blueprint-dash.bak/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
itsdangerous==2.2.0
google-cloud-bigquery==3.25.0
pandas==2.2.3
db-dtypes==1.3.1
google-api-python-client==2.149.0
google-auth==2.35.0
cachetools==5.5.0
python-dotenv==1.0.1
